add aes encryption of messages
A json envelope is added to encrypted messages with two keys: cipher and
payload. The payload is prepended with the iv of the encrypting cipher.

Payloads are currently not compatible with the openssl command line
client. It is not intended for users to set shared keys themselves,
rather they are generated by the server.