add api responder
Rather than have the api service talk to redis directly, implement the
api as a responder in the server. The api can talk to the server through
zeromq and let the server worry about where the responses are coming from.

This also makes it easy for clients to use the api. For example, plugins
could use stashes as memory.