call config callbacks on creation if key exists
If the key is already set, the callback won't get called because it
missed the initial broadcast. In most cases the callback should be
manually called with the current value. Pass the initial_set: false to
avoid this.

This is only enabled for callbacks that subscribe to a specific key and the
:set action.