subscribe before initial_set
There is a race condition between the initial_set and subscribe. With
the set first, you risk losing updates. With the subscribe first, you
risk having callbacks fire twice.

Callbacks firing twice is better than not at all, so the subscribe is
first. However this means callbacks should be idempotent.