make Zensu.node a registry
Serves to configure attributes of the local node. For now, that is
mostly the signing_key (ed25519).

LocalNode is a Registry (actually a Configuration) with special behavior
for the signing_key config var. When the signing_key is set, the id is
set to the signing_key's verify key.

Using verify keys for node ids allows any message to be signed and
verified by all participants.